gradient-service-inference/requirements/requirements.txt
attrs==21.2.0
bidict==0.21.2
black==21.9b0
certifi==2021.10.8
cffi==1.14.6
click==8.0.3
cloudpickle==1.6.0
coverage==6.0.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
future==0.18.2
gradient-service-domain==0.1.0+6
grpcio==1.33.2
hypothesis==6.0.0
iniconfig==1.1.1
invoke==1.6.0
kiwisolver==1.3.2
matplotlib==3.4.3
more-itertools==8.10.0
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.2
olefile==0.46
packaging==21.0
pandas==1.1.5
pathspec==0.9.0
Pillow==8.3.2
pip==21.3
platformdirs==2.3.0
pluggy==0.13.1
protobuf==3.18.1
py==1.10.0
pycparser==2.20
pymongo==3.11.0
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.10.8
scipy==1.7.1
setuptools==58.2.0
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
torch==1.9.1.post3
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
wheel==0.37.0

